remove unnecessary boxing and unboxing of ints in JEDI implementation
The lists storing the sorted aggregate sizes and children in a Node are
IntLists (which themselves exten List<Integer>). JEDIEvaluator calls
their get() functions which in turn box an int value into an Integer
which is immediately unboxed again which results in a high memory and
CPU time footprint. Calling getInt() instead fixes this problem
providing a speedup of 2-3x for a jedi() calculation.